test: move shared integration-test helper to tests/common/
Relocate tests/test_base.rs to tests/common/mod.rs so Cargo stops
compiling the helper as an empty test target, drop the
allow(unused_imports) workaround by importing Value directly where
needed, and dedupe the Value/artifact constructors shared by the
eval/statements.rs and eval/values.rs unit tests into a cfg(test)
eval::test_support module.

Closes #502
